+++ protocols/transport/tcp.c
@@ -1,6 +1,5 @@
 #include <stdio.h>
 #include <string.h>
-#include <arpa/inet.h>
 
 #include "tcp.h"
 #include "../../utils/visualizer.h"
@@ -63,7 +62,7 @@ void print_tcp_options(const uint8_t *pkt) {
             switch (kind) {
             case 2: {
                 if (length == 4) {
-                    mss = (((uint16_t)options[i + 2] << 8) | (uint16_t)options[i + 3]);
+                    mss = (uint16_t)(((uint16_t)options[i + 2] << 8) | (uint16_t)options[i + 3]);
                     printf("mss: %d", mss);
                 }
                 break;
@@ -83,14 +82,14 @@ void print_tcp_options(const uint8_t *pkt) {
                 printf("sack");
                 for (j = 2; j < length; j += 8) {
                     if (j < length - 7) {
-                        left_edge = ((uint32_t)options[i + j] << 24) |
+                        left_edge = (uint32_t)(((uint32_t)options[i + j] << 24) |
                                     ((uint32_t)options[i + j + 1] << 16) |
                                     ((uint32_t)options[i + j + 2] << 8)  |
-                                    ((uint32_t)options[i + j + 3]);
-                        right_edge = ((uint32_t)options[i + j + 4] << 24) |
+                                    ((uint32_t)options[i + j + 3]));
+                        right_edge = (uint32_t)(((uint32_t)options[i + j + 4] << 24) |
                                     ((uint32_t)options[i + j + 5] << 16) |
                                     ((uint32_t)options[i + j + 6] << 8)  |
-                                    ((uint32_t)options[i + j + 7]);
+                                    ((uint32_t)options[i + j + 7]));
                         printf(" (left_edge: %u, right_edge: %u)", left_edge, right_edge);
                     }
                 }
@@ -98,14 +97,14 @@ void print_tcp_options(const uint8_t *pkt) {
             }
             case 8: {
                 if (length == 10) {
-                    ts_val = ((uint32_t)options[i + 2] << 24) |
+                    ts_val = (uint32_t)(((uint32_t)options[i + 2] << 24) |
                             ((uint32_t)options[i + 3] << 16) |
                             ((uint32_t)options[i + 4] << 8)  |
-                            ((uint32_t)options[i + 5]);
-                    ts_echo = ((uint32_t)options[i + 6] << 24) |
+                            ((uint32_t)options[i + 5]));
+                    ts_echo = (uint32_t)(((uint32_t)options[i + 6] << 24) |
                             ((uint32_t)options[i + 7] << 16) |
                             ((uint32_t)options[i + 8] << 8)  |
-                            ((uint32_t)options[i + 9]);
+                            ((uint32_t)options[i + 9]));
                     printf("TSval: %u, TSecr: %u", ts_val, ts_echo);
                 }
                 break;
